Miscellaneous code cleanup
Biggest change is _capturesInternal, which no longer takes a slicer
argument.  The idea was that we could always thinSlice when we use it as
a Setter, that is, when it's guaranteed we won't be Getting out captures
and causing space leaks.  However, making this guarantee at the type level
is way out of scope for this project.